\name{irisimages}

\alias{irisimages}

\title{Iris images}

\description{
Set of iris images.
}

\usage{
irisimages
}

\section{Format}{
  Set of 3 iris JPEG images, in raw binary format.
}

\details{	
 These JPEG images are used in \code{hwriter} examples.
}

\seealso{
\code{hwriter}, \code{hwriteImage}
}

\author{
  Gregoire Pau, \email{gpau@ebi.ac.uk}, 2008
}

\keyword{interface}
